+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add `SystemSchema.hasChannel()` helper method.
+
 ### [v6.160](https://github.com/replicant4j/replicant/tree/v6.160) (2025-10-15) · [Full Changelog](https://github.com/replicant4j/replicant/compare/v6.159...v6.160)
 
 Changes in this release: